requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
numpy>=1.24.0
pytest>=7.0.0
